Refactor to allow merge with autora
