+++ autora/theorist/bsr/misc.py
